Moving community event handles to notification channel
